requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-decouple==3.8
django-ratelimit==4.1.0
django-filter==24.2
django-environ==0.12.0
whitenoise==6.6.0
cryptography==42.0.0

